refactor: move queue, publisher, subscriber to subpackages
