+++ subscriber/job.go
@@ -1,4 +1,4 @@
-package goeventqueue
+package subscriber
 
 import (
 	"context"
