+++ subscriber/subscriber.go
@@ -1,14 +1,17 @@
-package goeventqueue
+package subscriber
 
 import (
 	"context"
 	"fmt"
 	"sync"
+
+	"github.com/duysmile/goeventqueue"
+	"github.com/duysmile/goeventqueue/queue"
 )
 
 type Subscriber interface {
 	Start(ctx context.Context)
-	Register(name EventName, handler Handler)
+	Register(name goeventqueue.EventName, handler Handler)
 	WithLogger(l Logger)
 }
 
@@ -20,15 +23,15 @@ type Config struct {
 type Handler func(ctx context.Context, data interface{}) error
 
 type subscriber struct {
-	queue           Queue
-	mapEventHandler map[EventName][]Handler
+	queue           queue.Queue
+	mapEventHandler map[goeventqueue.EventName][]Handler
 	config          Config
 	locker          sync.Mutex
 	logger          Logger
 	quit            chan struct{}
 }
 
-func (s *subscriber) Register(name EventName, handler Handler) {
+func (s *subscriber) Register(name goeventqueue.EventName, handler Handler) {
 	s.locker.Lock()
 	defer s.locker.Unlock()
 
@@ -50,7 +53,7 @@ func (s *subscriber) Start(ctx context.Context) {
 	}
 }
 
-func (s *subscriber) startWorker(ctx context.Context, eQueue chan Event) {
+func (s *subscriber) startWorker(ctx context.Context, eQueue chan goeventqueue.Event) {
 	defer s.Recover()
 	for {
 		select {
@@ -91,10 +94,10 @@ func (s *subscriber) startWorker(ctx context.Context, eQueue chan Event) {
 	}
 }
 
-func NewSubscriber(q Queue, cfg Config) Subscriber {
+func NewSubscriber(q queue.Queue, cfg Config) Subscriber {
 	return &subscriber{
 		queue:           q,
-		mapEventHandler: make(map[EventName][]Handler),
+		mapEventHandler: make(map[goeventqueue.EventName][]Handler),
 		config:          cfg,
 		logger:          NewDefaultLogger(),
 	}
